Refactor creds controller to support refresh endpoint
Adds a refresh endpoint for the credentials report, to support slower
refresh times. This is to address rate-limit errors we are receiving
from AWS for these calls.

The IAM stuff was in the HQ controller for legacy reasons, it has been
migrated to a new controller. This means the stuff in the Utility
controller would be better placed in the now-sparse HQController, and
calling things "utils" is not ideal.
